requirements.txt
aiohttp==1.0.5
async-timeout==3.0.1
chardet==3.0.4
discord.py==0.16.12
multidict==4.5.2
selenium==3.141.0
urllib3==1.24.1
websockets==3.4
